Fix API edge cases:
When the API returned one page of results the next_url was missing, causing an
exception. We now check there's more than one page before we start looping.

If a space has been deleted there was an exception. We now catch this error.

There is still a bug in that the script incorectly reports that the org doesn't
exist when this happens.